+++ NEWS.rst
@@ -38,6 +38,9 @@ Misc:
   * Added ``--dry-run`` option to ``hiedr2mosaic.py`` (:numref:`hiedr2mosaic`).
   * Fixed a bug with very long input image names resulting in colliding
     ``.match`` and ``.vwip`` file names (:numref:`match_file_naming`).
+  * Added ``--mapproj-geolocation-uncertainty``, to keep stereo from failing on
+    mapprojected images with some misregistration and very little overlap
+    (:numref:`stereodefault-pprc`).
 
 RELEASE 3.7.0, June 2026
 ------------------------